F5B with F5C optimization and matrix reduction
A version of F5B with an optimization from F5C by Eder, Perry to compute using reduced GBs, which reduces the number of zero reductions, and with F4-style matrix reduction with an optional (but usually required) degree bound. Does not terminate reliably without the degree bound as yet, but produces correct results when it does as compared to F5B incremental. Matrices currently working with sympys built in matrix type and rref() function as proof of concept. Working on moving the same into numpy ndarrays and then adding numba cpu and gpu jit or vectorize decorators. Will attempt to write a PyCUDA SourceModule for more fine control if time allows. Does not interfere with cuda_cp or cuda_spoly, and will be able to interoperate with them.