Seems to be working pretty well
Working on adding a second version that launches 2 kernels to do both the multiply and subtract steps, but they need to be copied back to the host memory to be reindexed. That one isn't done yet. import cuda_spol and call cuda_s_poly(cp, ring). The symbolic preprocessing part has a lot of steps and so will probably be pretty slow, but its on the host side, so it might be faster than the second version that does each step on the gpu.