parallelized cuda cp for 32003 and 65521
Each of the files depends on a corresponding mul_inv_* file that is a protocol 4 pickled dictionary with all respective modular inverses pre calculated. They currently need to be sitting in the same directory but they could be moved to another directory as long as the path on the call to open() is changed. They both exist as global objects, and are added to the destination arrays for parsing after the kernel is finished.